+++ pkg/state/scale/consul/consul.go
@@ -6,6 +6,7 @@ import (
 	"strings"
 	"time"
 
+	"github.com/armon/go-metrics"
 	"github.com/gofrs/uuid"
 	"github.com/hashicorp/consul/api"
 	"github.com/jrasell/sherpa/pkg/client"
@@ -21,6 +22,16 @@ const (
 	baseKVPath         = "state/"
 	eventsKVPath       = "state/events/"
 	latestEventsKVPath = "state/latest-events/"
+)
+
+// Define our metric keys.
+var (
+	metricKeyGetEvents       = []string{"scale", "state", "consul", "get_events"}
+	metricKeyGetEvent        = []string{"scale", "state", "consul", "get_event"}
+	metricKeyGetLatestEvents = []string{"scale", "state", "consul", "get_latest_events"}
+	metricKeyGetLatestEvent  = []string{"scale", "state", "consul", "get_latest_event"}
+	metricKeyPutEvent        = []string{"scale", "state", "consul", "put_event"}
+	metricKeyGC              = []string{"scale", "state", "consul", "gc"}
 )
 
 type StateBackend struct {
@@ -47,6 +58,8 @@ func NewStateBackend(log zerolog.Logger, path string) scale.Backend {
 }
 
 func (s StateBackend) GetLatestScalingEvents() (map[string]*state.ScalingEvent, error) {
+	defer metrics.MeasureSince(metricKeyGetLatestEvents, time.Now())
+
 	kv, _, err := s.kv.List(s.latestEventsPath, nil)
 	if err != nil {
 		return nil, err
@@ -73,6 +86,8 @@ func (s StateBackend) GetLatestScalingEvents() (map[string]*state.ScalingEvent, 
 }
 
 func (s StateBackend) GetLatestScalingEvent(job, group string) (*state.ScalingEvent, error) {
+	defer metrics.MeasureSince(metricKeyGetLatestEvent, time.Now())
+
 	kv, _, err := s.kv.Get(s.latestEventsPath+job+":"+group, nil)
 	if err != nil {
 		return nil, err
@@ -90,6 +105,8 @@ func (s StateBackend) GetLatestScalingEvent(job, group string) (*state.ScalingEv
 }
 
 func (s StateBackend) GetScalingEvents() (map[uuid.UUID]map[string]*state.ScalingEvent, error) {
+	defer metrics.MeasureSince(metricKeyGetEvents, time.Now())
+
 	kv, _, err := s.kv.List(s.eventsPath, nil)
 	if err != nil {
 		return nil, err
@@ -122,6 +139,8 @@ func (s StateBackend) GetScalingEvents() (map[uuid.UUID]map[string]*state.Scalin
 }
 
 func (s StateBackend) GetScalingEvent(id uuid.UUID) (map[string]*state.ScalingEvent, error) {
+	defer metrics.MeasureSince(metricKeyGetEvent, time.Now())
+
 	kv, _, err := s.kv.List(s.eventsPath+id.String(), nil)
 	if err != nil {
 		return nil, err
@@ -147,6 +166,7 @@ func (s StateBackend) GetScalingEvent(id uuid.UUID) (map[string]*state.ScalingEv
 }
 
 func (s StateBackend) PutScalingEvent(job string, event *state.ScalingEventMessage) error {
+	defer metrics.MeasureSince(metricKeyPutEvent, time.Now())
 
 	sEntry := &state.ScalingEvent{
 		EvalID:  event.EvalID,
@@ -181,6 +201,9 @@ func (s StateBackend) PutScalingEvent(job string, event *state.ScalingEventMessa
 }
 
 func (s StateBackend) RunGarbageCollection() {
+	t := time.Now()
+	defer metrics.MeasureSince(metricKeyGC, t)
+
 	kv, _, err := s.kv.List(s.eventsPath, nil)
 	if err != nil {
 		s.logger.Error().Err(err).Msg("GC failed to list events in backend store")
@@ -190,7 +213,7 @@ func (s StateBackend) RunGarbageCollection() {
 		return
 	}
 
-	gc := time.Now().UTC().UnixNano() - s.gcThreshold
+	gc := t.UTC().UnixNano() - s.gcThreshold
 
 	for i := range kv {