+++ pkg/state/scale/memory/memory.go
@@ -4,12 +4,23 @@ import (
 	"sync"
 	"time"
 
+	"github.com/armon/go-metrics"
 	"github.com/gofrs/uuid"
 	"github.com/jrasell/sherpa/pkg/state"
 	"github.com/jrasell/sherpa/pkg/state/scale"
 )
 
 var _ scale.Backend = (*StateBackend)(nil)
+
+// Define our metric keys.
+var (
+	metricKeyGetEvents       = []string{"scale", "state", "memory", "get_events"}
+	metricKeyGetEvent        = []string{"scale", "state", "memory", "get_event"}
+	metricKeyGetLatestEvents = []string{"scale", "state", "memory", "get_latest_events"}
+	metricKeyGetLatestEvent  = []string{"scale", "state", "memory", "get_latest_event"}
+	metricKeyPutEvent        = []string{"scale", "state", "memory", "put_event"}
+	metricKeyGC              = []string{"scale", "state", "memory", "gc"}
+)
 
 type StateBackend struct {
 	gcThreshold int64
@@ -28,6 +39,8 @@ func NewStateBackend() scale.Backend {
 }
 
 func (s *StateBackend) GetLatestScalingEvents() (map[string]*state.ScalingEvent, error) {
+	defer metrics.MeasureSince(metricKeyGetLatestEvents, time.Now())
+
 	s.RLock()
 	latest := s.state.LatestEvents
 	s.RUnlock()
@@ -35,6 +48,8 @@ func (s *StateBackend) GetLatestScalingEvents() (map[string]*state.ScalingEvent,
 }
 
 func (s *StateBackend) GetLatestScalingEvent(job, group string) (*state.ScalingEvent, error) {
+	defer metrics.MeasureSince(metricKeyGetLatestEvent, time.Now())
+
 	s.RLock()
 	latest := s.state.LatestEvents[job+":"+group]
 	s.RUnlock()
@@ -42,6 +57,8 @@ func (s *StateBackend) GetLatestScalingEvent(job, group string) (*state.ScalingE
 }
 
 func (s *StateBackend) GetScalingEvents() (map[uuid.UUID]map[string]*state.ScalingEvent, error) {
+	defer metrics.MeasureSince(metricKeyGetEvents, time.Now())
+
 	s.RLock()
 	events := s.state.Events
 	s.RUnlock()
@@ -49,6 +66,8 @@ func (s *StateBackend) GetScalingEvents() (map[uuid.UUID]map[string]*state.Scali
 }
 
 func (s *StateBackend) PutScalingEvent(job string, event *state.ScalingEventMessage) error {
+	defer metrics.MeasureSince(metricKeyPutEvent, time.Now())
+
 	s.Lock()
 	defer s.Unlock()
 
@@ -71,6 +90,8 @@ func (s *StateBackend) PutScalingEvent(job string, event *state.ScalingEventMess
 }
 
 func (s *StateBackend) GetScalingEvent(id uuid.UUID) (map[string]*state.ScalingEvent, error) {
+	defer metrics.MeasureSince(metricKeyGetEvent, time.Now())
+
 	s.RLock()
 	e := s.state.Events[id]
 	s.RUnlock()
@@ -78,8 +99,10 @@ func (s *StateBackend) GetScalingEvent(id uuid.UUID) (map[string]*state.ScalingE
 }
 
 func (s *StateBackend) RunGarbageCollection() {
+	t := time.Now()
+	defer metrics.MeasureSince(metricKeyGC, t)
 
-	gc := time.Now().UTC().UnixNano() - s.gcThreshold
+	gc := t.UTC().UnixNano() - s.gcThreshold
 
 	newEventState := make(map[uuid.UUID]map[string]*state.ScalingEvent)